+++ src/app/app.component.ts
@@ -4,14 +4,14 @@ import {TranslateService} from '@ngx-translate/core';
 @Component({
   selector: 'free-root',
   template: `
-    <div class="free-main">
+    <div class="free-main" [class.free-main-mini]="mini">
       <router-outlet></router-outlet>
     </div>
   `,
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-
+  public mini: boolean;
   constructor(translate: TranslateService) {
     translate.setDefaultLang('en');
     translate.use('en');