+++ src/app/component/progress/progress.component.ts
@@ -5,8 +5,8 @@ import { NgModule, Component, AfterViewInit, Renderer2,
 @Component({
   selector: 'free-progress',
   template: `
-    <div class="free-progress" #container>
-      <div class="free-progress-bar" #bar>
+    <div class="free-progress" #container [ngStyle]="style" [style.background]="trackColor">
+      <div class="free-progress-bar" #bar [style.background]="barColor">
         <ng-container *ngIf="percent">{{value}}</ng-container>
       </div>
     </div>
@@ -19,15 +19,25 @@ export class ProgressComponent implements AfterViewInit {
   @Input() move: boolean;
   @Input() striped: boolean;
   @Input() percent: boolean;
+  @Input() width: any;
+  @Input() height: any;
+  @Input() barColor: string;
+  @Input() trackColor: string;
+  @Input() style: any;
   @ViewChild('container') container: ElementRef;
   @ViewChild('bar') bar: ElementRef;
   _bar: HTMLElement;
   _container: HTMLElement;
-  constructor(public renderer2: Renderer2) { }
+  constructor(public renderer2: Renderer2, public er: ElementRef) {
+    this.width = '200px';
+    this.height = '20px';
+  }
 
   ngAfterViewInit() {
     this._container = this.container.nativeElement;
     this._bar = this.bar.nativeElement;
+    this.renderer2.setStyle(this.er.nativeElement, 'width', this.width);
+    this.renderer2.setStyle(this.er.nativeElement, 'height', this.height);
 
     if (this.move) {
       this.renderer2.addClass(this._container, 'free-active');