+++ src/app/component/timeline/timeline.component.ts
@@ -1,7 +1,73 @@
-import {AfterViewInit, Component, ElementRef, Input, NgModule, OnInit, ViewChild} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, forwardRef, Inject, Input, NgModule, ViewChild} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {ShareModule} from '../common/share';
 import {DomRenderer} from '../common/dom';
+
+@Component({
+  selector: 'free-timeline-node',
+  template: `
+    <li class="free-timeline-node">
+      <ng-content></ng-content>
+    </li>
+  `
+})
+
+export class TimelineNodeComponent {
+}
+
+@Component({
+  selector: 'free-timeline-list',
+  template: `
+    <div class="free-timeline-list" [class.free-timeline-list-inverted]="inverted">
+      <ng-content></ng-content>
+    </div>
+  `
+})
+export class TimelineListComponent {
+  @Input() inverted: boolean;
+  constructor(@Inject(forwardRef(() => TimelineComponent)) timeline: TimelineComponent) {
+    timeline.line = true;
+  }
+}
+
+@Component({
+  selector: 'free-timeline-datetime',
+  template: `
+    <div class="free-timeline-datetime">
+        <span class="free-timeline-time">{{time}}</span>
+      <span class="free-timeline-date">{{date}}</span>
+    </div>
+  `
+})
+
+export class TimelineDatetimeComponent {
+  @Input() time: string;
+  @Input() date: string;
+}
+
+@Component({
+  selector: 'free-timeline-badge',
+  template: `
+    <div class="free-timeline-badge">
+      <ng-content></ng-content>
+    </div>
+  `
+})
+
+export class TimelineBadgeComponent {
+}
+
+@Component({
+  selector: 'free-timeline-body',
+  template: `
+    <div class="free-timeline-panel">
+      <ng-content></ng-content>
+    </div>
+  `
+})
+
+export class TimelineBodyComponent {
+}
 
 @Component({
   selector: 'free-timeline-item',
@@ -19,7 +85,9 @@ export class TimelineItemComponent implements AfterViewInit {
   @Input() dot: string;
   @ViewChild('item') itemViewChild: ElementRef;
   item: HTMLDivElement;
-  constructor(public domRenderer: DomRenderer) {}
+
+  constructor(public domRenderer: DomRenderer) {
+  }
 
   ngAfterViewInit() {
     this.item = this.itemViewChild.nativeElement;
@@ -34,22 +102,36 @@ export class TimelineItemComponent implements AfterViewInit {
 @Component({
   selector: 'free-timeline',
   template: `
-    <div class="free-timeline">
+    <div class="free-timeline" [class.free-timeline-line]="line">
       <ng-content></ng-content>
     </div>
   `
 })
-export class TimelineComponent implements OnInit {
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+export class TimelineComponent {
+  line: boolean;
 }
 
 @NgModule({
   imports: [CommonModule],
-  declarations: [TimelineItemComponent, TimelineComponent],
-  exports: [TimelineItemComponent, TimelineComponent, ShareModule]
+  declarations: [
+    TimelineItemComponent,
+    TimelineComponent,
+    TimelineNodeComponent,
+    TimelineListComponent,
+    TimelineDatetimeComponent,
+    TimelineBadgeComponent,
+    TimelineBodyComponent
+  ],
+  exports: [
+    TimelineItemComponent,
+    TimelineComponent,
+    TimelineNodeComponent,
+    TimelineListComponent,
+    ShareModule,
+    TimelineDatetimeComponent,
+    TimelineBadgeComponent,
+    TimelineBodyComponent
+  ]
 })
-export class TimelineModule {}
+export class TimelineModule {
+}