+++ src/app/democase/index/index-routing.module.ts
@@ -58,6 +58,8 @@ const childRoutes: Routes = [
   {path: 'cropper', loadChildren: '../main-cropper/main-cropper.module#MainCropperModule'},
   {path: 'panel', loadChildren: '../main-panel/main-panel.module#MainPanelModule'},
   {path: 'share-button', loadChildren: '../main-share-button/main-share-button.module#MainShareButtonModule'},
+  {path: 'steps', loadChildren: '../main-steps/main-steps.module#MainStepsModule'},
+  {path: 'knob', loadChildren: '../main-knob/main-knob.module#MainKnobModule'},
   {path: '', redirectTo: '/main/introduction', pathMatch: 'full'},
   {path: '**', redirectTo: '/error'}
 ];