+++ src/app/democase/main-knob/main-knob-routing.module.ts
@@ -1,0 +1,13 @@
+
+import {RouterModule} from '@angular/router';
+import {NgModule} from '@angular/core';
+import {MainKnobComponent} from './main-knob.component';
+
+@NgModule({
+  imports: [RouterModule.forChild([
+    {path: '', component: MainKnobComponent}
+  ])],
+  exports: [RouterModule]
+})
+
+export class MainKnobRoutingModule {}